requirements.txt
black==22.12.0
click==8.1.3
deprecation==2.1.0
lxml==4.9.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
pdf2image==1.16.2
pikepdf==6.2.8.post1
Pillow==9.4.0
platformdirs==2.6.2
reportlab==3.6.12
tomli==2.0.1
